old_junk/Database/Containers/honeybea/requirements.txt
fastapi==0.95.0
pydantic==1.10.4
PyMySQL==1.0.2
PyYAML==6.0
uvicorn>=0.15.0
